webapp(notifs): decouple notif handlers from invasion notifs
